+++ database/migrations/2020_04_06_014059_create_post_tag_table.php
@@ -16,8 +16,8 @@ class CreatePostTagTable extends Migration
         Schema::create('post_tag', function (Blueprint $table) {
             $table->bigIncrements('id');
 
-            $table->integer('post_id')->unsigned;
-            $table->integer('tag_id')->unsigned;
+            $table->bigInteger('post_id')->unsigned;
+            $table->bigInteger('tag_id')->unsigned;
 
             $table->timestamps();
